fix(ci): fix markdown lint, Netlify build, and split oversized test file
